web/rest: Avoid API class duplications
