+++ library/Ivoz/Provider/Infrastructure/Api/Jwt/UserAuthenticationSuccessHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jwt;
+namespace Ivoz\Provider\Infrastructure\Api\Jwt;
 
 use Symfony\Component\Security\Core\User\UserInterface;
 use Lexik\Bundle\JWTAuthenticationBundle\Security\Http\Authentication\AuthenticationSuccessHandler as OriginalAuthenticationSuccessHandler;
