+++ web/rest/brand/app/config/security.yml
@@ -19,9 +19,9 @@ security:
 
   providers:
     admin_orm:
-      id: 'Security\User\Provider::Administrator'
+      id: 'Ivoz\Provider\Infrastructure\Api\Security\User\Provider::Administrator'
     user_orm:
-      id: 'Security\User\Provider::User'
+      id: 'Ivoz\Provider\Infrastructure\Api\Security\User\Provider::User'
 
   firewalls:
     # disables authentication for assets and the profiler, adapt it according to your needs
@@ -57,7 +57,7 @@ security:
           require_previous_session: false
           username_parameter: email
           password_parameter: password
-          success_handler: Jwt\UserAuthenticationSuccessHandler
+          success_handler: Ivoz\Provider\Infrastructure\Api\Jwt\UserAuthenticationSuccessHandler
           failure_handler: lexik_jwt_authentication.handler.authentication_failure
           require_previous_session: false
 
@@ -71,7 +71,7 @@ security:
       stateless: true
       guard:
         authenticators:
-          - Jwt\TokenAuthenticator
+          - Ivoz\Provider\Infrastructure\Api\Jwt\TokenAuthenticator
 
     main:
       anonymous: ~
