+++ web/rest/brand/app/config/services.yml
@@ -9,25 +9,9 @@ services:
     autoconfigure: true
     public: false
 
-  Jwt\UserAuthenticationSuccessHandler:
+  Service\AuthEndpointDecorator:
+    decorates: 'api_platform.swagger.normalizer.documentation'
+    decoration_priority: 5
     arguments:
-      - '@lexik_jwt_authentication.jwt_manager'
-      - '@event_dispatcher'
-
-  Jwt\TokenAuthenticator:
-    arguments:
-      - '@lexik_jwt_authentication.jwt_manager'
-      - '@event_dispatcher'
-      - '@lexik_jwt_authentication.extractor.chain_extractor'
-
-  Security\User\Provider::User:
-    class: Security\User\UserProvider
-    arguments:
-      $identifierField: 'email'
-      $entityClass: 'Ivoz\Provider\Domain\Model\User\User'
-
-  Security\User\Provider::Administrator:
-    class: Security\User\UserProvider
-    arguments:
-      $identifierField: 'username'
-      $entityClass: 'Ivoz\Provider\Domain\Model\Administrator\Administrator'
+      $decoratedNormalizer:
+        '@Service\AuthEndpointDecorator.inner'
